+++ appwidget/src/main/java/jp/numero/dagashiapp/appwidget/WidgetStyle.kt
